Emit error if upload is locked and allow retries
Previously, if an upload was locked, the client would create a new upload and would therefore restart
from the begining. An upload may be locked for a short time if the connection got interrupted and the
server didn't have time to clean up the dead connection. Therefore, the upload will usually be
available soon to resume. By emitting an error, the retry mechanism can catch this situation and
retry accordingly.

Closes #37