Emit an initial progress event.
This patch emits a progress event as soon as the upload starts. When
resuming, this means you get the current upload offset as soon as
possible. When starting a new upload, this allows you to listen for the
`onProgress` event if you need to get the `upload.url` as soon as
possible, when not using fingerprinting.