Added support for Cordova File
Since a File object loaded with the Cordova file plugin is not exactly the same a one loaded with a file input some changes were needed to support them.

The slice method of a Cordova File returns a new File object which needs to be loaded with a FileReader. Since this is an async operation, I've implemented a promise handler to accomodate.