Do not invoke callbacks for aborted Node.js request
Previously, an upload would be retried if an upload with retryDelays
is aborted. It is caused by the fact that an error is emitted when the
HTTP request is aborted. This error is handled by the retry mechanism
which then retries the upload. However, the user does not want to retry
if they abort the upload.

Fixes https://github.com/tus/tus-js-client/issues/138