Disable babel-preset-env `useBuiltIns` option
`useBuiltIns` injects loads of polyfills that are installed into the
environment. core-js polyfills are very thorough, but that sometimes
means that it includes things that are actually totally unnecessary.

For example, tus-js-client@2 shipped with a Symbol polyfill, even though
it doesn't use Symbols anywhere. Transpiled array rest spread syntax uses
`Symbol.iterator`, though, so preset-env figures that `Symbol` is in use,
and adds the polyfill. This requires core-js to override things like
`Object.getOwnPropertyNames` to exclude polyfilled Symbols, etc, and it
just spirals out of control until you have 50KB of polyfills just for
array spread syntax and Promises :)

The only feature that tus-js-client _actually_ appears to require is
`Promise`, which we can reasonably expect users to polyfill.