Use ES5 for `"module"` entry point, too
`"module"` and `"main"` are both commonly used by bundlers. To support
the widest range of browsers, they should be transpiled to ES5. It's
possible and not weird to use newer syntax, but best if people opt in
to it, which they can do with a resolve alias. In webpack for example:

```js
resolve: {
  alias: {
    'tus-js-client': path.join(__dirname, 'node_modules/tus-js-client/lib')
  }
}
```