Avoid situation where callback assertions don't work
See the problem in `test/spec/test-web-stream.js`. If `onSuccess` is a normal function, `onSuccess.toBeCalled` and `await onSuccess.toBeCalled` resolves immediately, without asserting that the callback was invoked.

While the fix is to make `onSuccess` a spy, similar situations can go unnoticed. To make this logic error out if the callback isn't a spy, this commit changes `toBeCalled` to be a function, so `await onSuccess.toBeCalled()` would error out if `onSuccess` is a normal function.