Fix deferred length retry behavior for final chunk upload request
Before this change, if the last chunk upload resulted in a retryable error, the
retry will not have the `Upload-Length` header set and the retry will
fail. Since there is no need to remove
`upload._uploadLengthDeferred` and doing so results in the header not
being set on the retry, this PR leaves it set.